Adding a little note about nested groups
As a Laravel user that came to Lumen, I kind of expected groups to work in the same way. I'm assuming that as an optimisation, the group stack was removed. It'd be nice to highlight this so that people (like myself) don't have to wade through the internals to find that it's a feature and not a bug. ;-)