Updates installation file
Being new to Lumen (and the PHP community) I found that starting a Lumen app was a bit more difficult than I expected. This is because the docs currently do not specify a way to start a Lumen app. After scouring the internet, wanting to find the right (lumen-specific) command, I finally came to the conclusion that Lumen apps seem to be spun up using the PHP server. While it may seem apparent, I think it would be really great to add a quick, explicit section on how to start a Lumen app! 💡