Add documentation for optional route parameters
nikic/FastRoute has support for optional route route parts and parameters. This is easily useable in Lumen so should be documented.
I've added an "Optional Parameters" section (similar to whats used in Laravel docs) but it could be better named since the optional URL parts can contain parameters, or just be static.